backend/tourstar.kroaddy.site/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
ultralytics>=8.3.0
pyiqa>=0.1.13
Pillow>=10.0.0
langgraph>=0.2.0


